voice notes of each entry will be saved locally and online

+++ app/src/main/java/com/ledgero/ViewModels/UnApprovedEntriesViewModel.kt
@@ -21,22 +21,45 @@ class UnApprovedEntriesViewModel(private val unApprovedEntriesRepo: UnApprovedEn
 
     }
 
-    fun deleteEntry(pos : Int){
-        unApprovedEntriesRepo.deleteEntry(pos)
-    }
     private fun getEntriesFromRepo(): LiveData<ArrayList<Entries>> {
 
         return unApprovedEntriesRepo.getEntries()
     }
 
-    fun removeListener(){
-        unApprovedEntriesRepo.removeListener()
-    }
+    //this function will be called when entry has to be deleted by requester
+    fun deleteEntry(pos : Int){
+        if (allUnApprovedEntries.value!!.get(pos).hasVoiceNote!!){
+            //deleteEntry from unapproved if it has voice note
+            unApprovedEntriesRepo.deleteEntry(allUnApprovedEntries.value!!.get(pos))
+        }else{
+            //deleteEntry from unapproved if it does not have voice note
+        unApprovedEntriesRepo.deleteEntry(pos)
+    }}
 
+
+    //this will be called when receiver has accepted requester entry add request
     fun approveEntry(pos:Int){
+        // we don't need to check if entry has voice not while approving the adding a new entry
+        // because we will be downloading voice if not present in device when user
+        // open each entry for edit/review
         unApprovedEntriesRepo.approveEntry(pos)
     }
 
+    //this function will be called when receiver has accepted the request to delete a entry from ledger
+    fun deleteEntryFromLedgerRequest_accepted(pos: Int){
+        //check if entry to be delete has voice not or not
+        var entry = allUnApprovedEntries.value!!.get(pos)
+
+        if(entry.hasVoiceNote!!){
+            unApprovedEntriesRepo.acceptDeleteEntryRequestFromApprovedEntries_withVoice(entry)
+        }else{
+            unApprovedEntriesRepo.deleteEntry(pos)
+        }
+
+
+    }
+
+    //this will be called when receiver has rejected the requester entry add/del request
     fun rejectEntry(pos:Int){
 
         //check if rejected Entry was new entry or was approved entry requested for delete/update
@@ -44,7 +67,7 @@ class UnApprovedEntriesViewModel(private val unApprovedEntriesRepo: UnApprovedEn
 
         if (entry.requestMode==1)//means it was just a new entry requested to add, so we can simply delete it from un-approved and move it to canceled
         {
-            unApprovedEntriesRepo.rejectNewAddedEntry(pos)
+            unApprovedEntriesRepo.rejectNewAddedEntry(entry)
         }
         if (entry.requestMode==2){
             // means it was a approved Entry requested to delete
@@ -53,12 +76,18 @@ class UnApprovedEntriesViewModel(private val unApprovedEntriesRepo: UnApprovedEn
         }
 
 
-        unApprovedEntriesRepo.rejectedEntry(pos)
+       // unApprovedEntriesRepo.rejectedEntry(entry)
     }
 
     fun deleteUnApprovedEntryThenUpdateLedgerEntry(pos: Int) {
         unApprovedEntriesRepo.deleteUnApprovedEntryThenUpdateLedgerEntry(pos)
 
+    }
+
+
+
+    fun removeListener(){
+        unApprovedEntriesRepo.removeListener()
     }
 
 }
